requirements.txt
APScheduler==3.6.3
astroid==2.8.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
loguru==0.5.3
mccabe==0.6.1
platformdirs==2.4.0
pydantic==1.8.2
pylint==2.11.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
sentry-sdk==1.4.2
six==1.16.0
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
wrapt==1.12.1
